test: add missing error handling tests for token store chmod
